generalize configurable link button
